+++ runtime/shared/fbt.js
@@ -13,45 +13,36 @@
  * Run the following command to sync the change from www to fbsource.
  *   js1 upgrade www-shared -p fbt --local ~/www
  *
- * @flow
+ * @flow strict-local
  * @typechecks
  * @format
  * @emails oncall+internationalization
  */
 
 /* eslint max-len: ["warn", 120] */
-/* eslint "fb-www/avoid-this-outside-classes": "off" */
-/* eslint "fb-www/order-requires": "off" */
-/* eslint "fb-www/require-flow-strict-local": "off" */
 require('FbtEnv').setup();
 
-const {overrides} = require('FbtQTOverrides');
+import type {FbtInputOpts, FbtRuntimeInput, FbtTableArgs} from 'FbtHooks';
+
 const FbtHooks = require('FbtHooks');
+const {overrides} = require('FbtQTOverrides');
 const FbtResultBase = require('FbtResultBase');
+const FbtTable = require('FbtTable');
 const FbtTableAccessor = require('FbtTableAccessor');
 const GenderConst = require('GenderConst');
+const {
+  getGenderVariations,
+  getNumberVariations,
+} = require('IntlVariationResolver');
 const IntlViewerContext = require('IntlViewerContext');
 
 const intlNumUtils = require('intlNumUtils');
 const invariant = require('invariant');
 const substituteTokens = require('substituteTokens');
-const {
-  getNumberVariations,
-  getGenderVariations,
-} = require('IntlVariationResolver');
-
-import type {FbtInputOpts, FbtRuntimeInput, FbtTableArgs} from 'FbtHooks';
 
 let jsonExportMode = false; // Used only in React Native
 
-/**
- * fbt.XXX calls return arguments in the form of
- * [<INDEX>, <SUBSTITUTION>] to be processed by fbt._
- */
-const ARG = {
-  INDEX: 0,
-  SUBSTITUTION: 1,
-};
+const {ARG} = FbtTable;
 
 const VARIATIONS = {
   NUMBER: 0,
@@ -141,7 +132,7 @@ fbt._ = function(
   // args that follow.
   let allSubstitutions = {};
 
-  if (pattern.__vcg) {
+  if (pattern.__vcg != null) {
     args = args || [];
     const {GENDER} = IntlViewerContext;
     const variation = getGenderVariations(GENDER);
@@ -152,7 +143,7 @@ fbt._ = function(
     if (typeof pattern !== 'string') {
       // On mobile, table can be accessed at the native layer when fetching
       // translations. If pattern is not a string here, table has not been accessed
-      pattern = _accessTable(pattern, args, 0);
+      pattern = FbtTable.access(pattern, args, 0);
     }
     allSubstitutions = Object.assign(
       {},
@@ -172,8 +163,8 @@ fbt._ = function(
     // (see intl_get_application_id)
     // $FlowFixMe pattern is the tuple [string, string]
     const stringID = '1_' + patternHash;
-    patternString = overrides[stringID] || patternString;
-    if (overrides[stringID]) {
+    if (overrides[stringID] != null && overrides[stringID] !== '') {
+      patternString = overrides[stringID];
       FbtHooks.onTranslationOverride(patternHash);
     }
     FbtHooks.logImpression(patternHash);
@@ -185,12 +176,12 @@ fbt._ = function(
   }
 
   const cachedFbt = _cachedFbtResults[patternString];
-  const hasSubstitutions = this._hasKeys(allSubstitutions);
+  const hasSubstitutions = _hasKeys(allSubstitutions);
   if (cachedFbt && !hasSubstitutions) {
     return cachedFbt;
   } else {
     const fbtContent = substituteTokens(patternString, allSubstitutions);
-    const result = this._wrapContent(fbtContent, patternString, patternHash);
+    const result = _wrapContent(fbtContent, patternString, patternHash);
     if (!hasSubstitutions) {
       _cachedFbtResults[patternString] = result;
     }
@@ -203,104 +194,15 @@ if (__DEV__) {
 }
 
 /**
- * fbt._hasKeys takes an object and returns whether it has 0
- * keys. It purposefully avoids creating the temporary arrays
- * incurred by calling Object.keys(o)
+ * _hasKeys takes an object and returns whether it has any keys. It purposefully
+ * avoids creating the temporary arrays incurred by calling Object.keys(o)
  * @param {Object} o - Example: "allSubstitutions"
  */
-
-fbt._hasKeys = function(o) {
+function _hasKeys(o) {
   for (const k in o) {
     return true;
   }
   return false;
-};
-
-/**
- * Performs a depth-first search on our table, attempting to access
- * each table entry.  The first entry found is the one we want, as we
- * set defaults after preferred indices.  For example:
- *
- * @param @table - {
- *   // viewer gender
- *   '*': {
- *     // {num} plural
- *     '*': {
- *       // user-defined enum
- *       LIKE: '{num} people liked your update',
- *       COMMENT: '{num} people commented on your update',
- *       POST: '{num} people posted on a wall',
- *     },
- *     SINGULAR: {
- *       LIKE: '{num} person liked your update',
- *       // ...
- *     },
- *     DUAL: { ... }
- *   },
- *   FEMALE: {
- *     // {num} plural
- *     '*': { ... },
- *     SINGULAR: { ... },
- *     DUAL: { ... }
- *   },
- *   MALE: { ... }
- * }
- *
- * Notice that LIKE and COMMENT here both have 'your' in them, whereas
- * POST doesn't.  The fallback ('*') translation for POST will be the same
- * in both the male and female version, so that entry won't exist under
- *   table[FEMALE]['*'] or table[MALE]['*'].
- *
- * Similarly, PLURAL is a number variation that never appears in the table as it
- * is the default/fallback.
- *
- * For example, if we have a female viewer, and a PLURAL number and a POST enum
- * value, in the above example, we'll first attempt to get:
- * table[FEMALE][PLURAL][POST].  undefined. Back Up, attempting to get
- * table[FEMALE]['*'][POST].  undefined also. since it's the same as the '*'
- * table['*'][PLURAL][POST].  ALSO undefined. Deduped to '*'
- * table['*']['*'][POST].  There it is.
- *
- * @param args      - fbt runtime arguments
- * @param argsIndex - argument index we're currently visiting
- */
-function _accessTable(
-  table: ?FbtRuntimeInput,
-  args: FbtTableArgs,
-  argsIndex: number,
-): ?string | [string, string] {
-  if (argsIndex >= args.length) {
-    // We've reached the end of our arguments at a valid entry, in which case
-    // table is now a string (leaf)
-    // $FlowFixMe string is incompatible FbtInputTable type
-    return table;
-  } else if (table == null) {
-    // We've accessed a key that didn't exist in the table
-    return null;
-  }
-  const arg = args[argsIndex];
-  const tableIndices = arg[ARG.INDEX];
-
-  if (tableIndices == null) {
-    return _accessTable(table, args, argsIndex + 1);
-  }
-  invariant(
-    typeof table !== 'string',
-    'If tableIndex is non-null, we should have a table, but we got: %s',
-    table,
-  );
-  // Is there a variation? Attempt table access in order of variation preference
-  for (let k = 0; k < tableIndices.length; ++k) {
-    const key = tableIndices[k];
-    // table isn't a tuple here, but flow thinks it could be
-    // $FlowFixMe string is not an array index
-    const subTable = table[key];
-    const pattern = _accessTable(subTable, args, argsIndex + 1);
-    if (pattern != null) {
-      return pattern;
-    }
-  }
-  return null;
 }
 
 /**
@@ -458,19 +360,21 @@ fbt._name = function(label, value, gender) {
   return FbtTableAccessor.getGenderResult(variation, substitution, gender);
 };
 
-fbt._wrapContent = (fbtContent, patternString, patternHash) => {
+function _wrapContent(fbtContent, patternString, patternHash): Fbt {
   const contents = typeof fbtContent === 'string' ? [fbtContent] : fbtContent;
   const errorListener = FbtHooks.getErrorListener({
     translation: patternString,
     hash: patternHash,
   });
-  return FbtHooks.getFbtResult({
+  const result = FbtHooks.getFbtResult({
     contents,
     errorListener,
     patternString,
     patternHash,
   });
-};
+  // $FlowFixMe FbtHooks.getFbtResult returns mixed.
+  return result;
+}
 
 fbt.enableJsonExportMode = function() {
   jsonExportMode = true;